refactor: rename questions_answers/ package to questions_and_answers/
Aligns the package directory name with the module file (questions_and_answers.py) and the class (QuestionsAndAnswers).
