+++ worker_plan/worker_plan_internal/plan/nodes/questions_and_answers.py
@@ -1,7 +1,7 @@
 """QuestionsAndAnswersTask - Generates Q&A about the plan."""
 from llama_index.core.llms.llm import LLM
 from worker_plan_internal.plan.run_plan_pipeline import PlanTask
-from worker_plan_internal.questions_answers.questions_and_answers import QuestionsAndAnswers
+from worker_plan_internal.questions_and_answers.questions_and_answers import QuestionsAndAnswers
 from worker_plan_api.filenames import FilenameEnum
 from worker_plan_internal.plan.nodes.strategic_decisions_markdown import StrategicDecisionsMarkdownTask
 from worker_plan_internal.plan.nodes.scenarios_markdown import ScenariosMarkdownTask
